[Vue3 + TS + Quasar] refactoring: Moved graphql setup functionality into a separate file
