+++ vue3-ts-quasar/src/init/index.ts
@@ -1,0 +1,5 @@
+/* 
+    This file is simply meant to be an aggregator of all init functions.
+    This file does not implement any logic of its own and simply serves the purpose of importing functions and exporting them out together so that they can be loaded from one file.
+*/
+export { setupGraphQL } from './setupGraphQL';
