+++ vue3-ts-quasar/src/main.ts
@@ -1,14 +1,6 @@
-import { createApp, provide, h } from 'vue';
+import { createApp, h } from 'vue';
 import App from './App.vue';
 import router from './router';
-
-// GraphQL related
-import {
-  ApolloClient,
-  createHttpLink,
-  InMemoryCache,
-} from '@apollo/client/core';
-import { DefaultApolloClient } from '@vue/apollo-composable';
 
 // Store - Global state management
 import { createPinia } from 'pinia';
@@ -17,27 +9,16 @@ import { createPinia } from 'pinia';
 import { Quasar } from 'quasar';
 import quasarUserOptions from './quasar-user-options';
 
+//
+import { setupGraphQL } from './init';
+
 const pinia = createPinia();
 
 //* GraphQL setup
-// HTTP connection to the API
-const httpLink = createHttpLink({
-  // You should use an absolute URL here
-  uri: process.env.VUE_APP_GRAPHQL_URL,
-});
-
-// Cache implementation
-const cache = new InMemoryCache();
-
-// Create the apollo client
-const apolloClient = new ApolloClient({
-  link: httpLink,
-  cache,
-});
 
 createApp({
   setup() {
-    provide(DefaultApolloClient, apolloClient);
+    setupGraphQL();
   },
   render: () => h(App),
 })
